move contents of commands/util.py to commands/signal.py
